[AN] feat: 최근 업로드된 히어릿 들은 위치 progressBar 표시 #629

+++ android/app/src/main/java/com/onair/hearit/data/dto/RecentUploadResponse.kt
@@ -12,7 +12,7 @@ data class RecentUploadResponse(
     @SerialName("playTime")
     val playTime: Int,
     @SerialName("lastPlayTime")
-    val lastPlayTime: Int,
+    val lastPlayTime: Long,
     @SerialName("createdAt")
     val createdAt: String,
     @SerialName("category")
